Создать mathFact.model, обновить mathFact.controller
Вся суть этого коммита в том, что мы делегировали запросы в базу данных
в mathFact.model. Было сделано это А) для поддержания тестируемости кода,
т.к создавать стаб на pool.query (pool: pg.Pool) было невозможно из-за
типизации, т.к по дефолту в query возвращается void. Б) Из-за новой идеи
для архитектуры приложения, которую я увидел
(контроллер-сервис-репозиторий). Впринципе идейка неплохая, почему бы и
не подспиздить это для своих проектиков.
